chore: refactor expression parser to avoid stack overflow issue
